phase-0-mvp/requirements.txt
# Smalltalk Crash Analyzer - MVP Dependencies
# Install with: pip install -r requirements.txt

# AI Providers
openai>=1.0.0
anthropic>=0.9.0

# Configuration
pyyaml>=6.0

# Optional: Local AI
# ollama>=0.1.0
